convert TermsOfServicePage to tsx and use RemoteMarkdown
